requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
Flask-Mail>=0.9.0
PyMySQL>=1.1.0
mysql-connector-python>=8.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
Pillow>=10.0.0
openpyxl>=3.1.0
pandas>=2.1.0
python-dateutil>=2.8.0
bcrypt>=4.0.0
email-validator>=2.0.0
